requirements.txt
# Core computer vision and ML libraries
opencv-python>=4.8.0
dlib>=19.24.0
numpy>=1.24.0
scikit-learn>=1.3.0
torch>=2.0.0
torchvision>=0.15.0

# Modern ML libraries
transformers>=4.30.0
datasets>=2.14.0
evaluate>=0.4.0

# Web interface
streamlit>=1.25.0
plotly>=5.15.0

# Data handling and utilities
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pillow>=10.0.0

# Configuration and logging
pyyaml>=6.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
